Modal

+++ DL_EF/Alumno.cs
@@ -20,7 +20,7 @@ namespace DL_EF
         public string ApellidoMaterno { get; set; }
         public Nullable<System.DateTime> FechaNacimiento { get; set; }
         public string Sexo { get; set; }
-        public Nullable<int> IdSemestre { get; set; }
+        public Nullable<byte> IdSemestre { get; set; }
     
         public virtual Semestre Semestre { get; set; }
     }
